feat!: new @vueuse/metadata package
`indexes.json` is no longer existed in `@vueuse/core`, instead, import from `@vueuse/metadata` or `@vueuse/core/metadata`
