+++ meta/packages.ts
@@ -1,6 +1,14 @@
-import type { PackageManifest } from './types'
+import type { PackageManifest } from '@vueuse/metadata'
 
 export const packages: PackageManifest[] = [
+  {
+    name: 'metadata',
+    display: 'Metadata for VueUse functions',
+    manualImport: true,
+    iife: false,
+    utils: true,
+    target: 'node14',
+  },
   {
     name: 'shared',
     display: 'Shared utilities',
@@ -28,6 +36,7 @@ export const packages: PackageManifest[] = [
     addon: true,
     iife: false,
     cjs: false,
+    utils: true,
     target: 'node14',
     external: [
       '@vueuse/core',
