+++ packages/.vitepress/plugins/markdownTransform.ts
@@ -2,7 +2,7 @@ import { join, resolve } from 'path'
 import type { Plugin } from 'vite'
 import fs from 'fs-extra'
 import { packages } from '../../../meta/packages'
-import { functionNames, getFunction } from '../../../meta/function-indexes'
+import { functionNames, getFunction } from '../../../packages/metadata/metadata'
 import { getTypeDefinition, replacer } from '../../../scripts/utils'
 
 export function MarkdownTransform(): Plugin {
