+++ packages/.vitepress/theme/components/Contributors.vue
@@ -2,7 +2,7 @@
 // @ts-expect-error missing types
 import _contributors from '/virtual-contributors'
 import { computed } from 'vue'
-import type { ContributorInfo } from '../../../../meta/types'
+import type { ContributorInfo } from '../@vueuse/metadata'
 import { renderCommitMessage } from '../utils'
 
 const props = defineProps<{ fn: string }>()
