+++ packages/core/metadata.cjs
@@ -1,0 +1,1 @@
+module.exports = require('@vueuse/metadata')
