+++ packages/core/metadata.d.ts
@@ -1,0 +1,1 @@
+export * from '@vueuse/metadata'
