+++ packages/core/metadata.mjs
@@ -1,0 +1,1 @@
+export * from '@vueuse/metadata'
