+++ packages/metadata/.gitignore
@@ -1,0 +1,1 @@
+index.json
