+++ scripts/changelog.ts
@@ -1,7 +1,7 @@
 import md5 from 'md5'
 import Git from 'simple-git'
-import type { CommitInfo, ContributorInfo } from '../meta/types'
-import { functions } from '../meta/function-indexes'
+import type { CommitInfo, ContributorInfo } from '@vueuse/metadata'
+import { functions } from '../packages/metadata/metadata'
 import { uniq } from './utils'
 
 const git = Git({
