+++ scripts/update.ts
@@ -1,16 +1,14 @@
 import fs from 'fs-extra'
-import { readIndexes, updateFunctionREADME, updateFunctionsMD, updateImport, updateIndexREADME, updatePackageJSON, updatePackageREADME } from './utils'
+import { metadata } from '../packages/metadata/metadata'
+import { updateFunctionREADME, updateFunctionsMD, updateImport, updateIndexREADME, updatePackageJSON, updatePackageREADME } from './utils'
 
 async function run() {
-  const indexes = await readIndexes()
-
-  fs.writeJSON('indexes.json', indexes, { spaces: 2 })
-  await updateImport(indexes)
-  await updatePackageREADME(indexes)
-  await updateIndexREADME(indexes)
-  await updateFunctionsMD(indexes)
-  await updateFunctionREADME(indexes)
-  await updatePackageJSON(indexes)
+  await updateImport(metadata)
+  await updatePackageREADME(metadata)
+  await updateIndexREADME(metadata)
+  await updateFunctionsMD(metadata)
+  await updateFunctionREADME(metadata)
+  await updatePackageJSON(metadata)
 
   await fs.copy('./CONTRIBUTING.md', './packages/contributing.md')
 }
