+++ scripts/utils.ts
@@ -1,21 +1,20 @@
-import { join, relative, resolve } from 'path'
+import { join, resolve } from 'path'
 import fs from 'fs-extra'
 import matter from 'gray-matter'
-import fg from 'fast-glob'
 import parser from 'prettier/parser-typescript'
 import prettier from 'prettier'
 import YAML from 'js-yaml'
 import Git from 'simple-git'
-import { ecosystemFunctions } from '../meta/ecosystem-functions'
+import type { PackageIndexes, VueUseFunction } from '@vueuse/metadata'
 import { packages } from '../meta/packages'
-import type { PackageIndexes, VueUseFunction, VueUsePackage } from '../meta/types'
+import { getCategories } from '../packages/metadata/utils'
 
-const git = Git()
+export const git = Git()
 
-const DOCS_URL = 'https://vueuse.org'
+export const DOCS_URL = 'https://vueuse.org'
 
-const DIR_ROOT = resolve(__dirname, '..')
-const DIR_SRC = resolve(__dirname, '../packages')
+export const DIR_ROOT = resolve(__dirname, '..')
+export const DIR_SRC = resolve(__dirname, '../packages')
 const DIR_TYPES = resolve(__dirname, '../types/packages')
 
 export async function getTypeDefinition(pkg: string, name: string): Promise<string | undefined> {
@@ -49,110 +48,6 @@ export async function getTypeDefinition(pkg: string, name: string): Promise<stri
 
 export function hasDemo(pkg: string, name: string) {
   return fs.existsSync(join(DIR_SRC, pkg, name, 'demo.vue'))
-}
-
-export async function listFunctions(dir: string, ignore: string[] = []) {
-  const files = await fg('*', {
-    onlyDirectories: true,
-    cwd: dir,
-    ignore: [
-      '_*',
-      'dist',
-      'node_modules',
-      ...ignore,
-    ],
-  })
-  files.sort()
-  return files
-}
-
-export async function readIndexes() {
-  const indexes: PackageIndexes = {
-    packages: {},
-    categories: [],
-    functions: [
-      ...ecosystemFunctions,
-    ],
-  }
-
-  for (const info of packages) {
-    const dir = join(DIR_SRC, info.name)
-
-    const functions = await listFunctions(dir)
-
-    const pkg: VueUsePackage = {
-      ...info,
-      dir: relative(DIR_ROOT, dir).replace(/\\/g, '/'),
-      docs: info.addon ? `${DOCS_URL}/${info.name}/README.html` : undefined,
-    }
-
-    indexes.packages[info.name] = pkg
-
-    await Promise.all(functions.map(async(fnName) => {
-      const mdPath = join(dir, fnName, 'index.md')
-      const tsPath = join(dir, fnName, 'index.ts')
-
-      const fn: VueUseFunction = {
-        name: fnName,
-        package: pkg.name,
-        lastUpdated: +await git.raw(['log', '-1', '--format=%at', tsPath]) * 1000,
-      }
-
-      if (fs.existsSync(join(dir, fnName, 'component.ts')))
-        fn.component = true
-      if (fs.existsSync(join(dir, fnName, 'directive.ts')))
-        fn.directive = true
-
-      if (!fs.existsSync(mdPath)) {
-        fn.internal = true
-        indexes.functions.push(fn)
-        return
-      }
-
-      fn.docs = `${DOCS_URL}/${pkg.name}/${fnName}/`
-
-      const mdRaw = await fs.readFile(mdPath, 'utf-8')
-
-      const { content: md, data: frontmatter } = matter(mdRaw)
-      const category = frontmatter.category
-
-      let description = (md
-        .replace(/\r\n/g, '\n')
-        .match(/# \w+[\s\n]+(.+?)(?:, |\. |\n|\.\n)/m) || []
-      )[1] || ''
-
-      description = description.trim()
-      description = description.charAt(0).toLowerCase() + description.slice(1)
-
-      fn.category = ['core', 'shared'].includes(pkg.name) ? category : `@${pkg.display}`
-      fn.description = description
-
-      if (description.includes('DEPRECATED'))
-        fn.deprecated = true
-
-      indexes.functions.push(fn)
-    }))
-  }
-
-  indexes.functions.sort((a, b) => a.name.localeCompare(b.name))
-  indexes.categories = getCategories(indexes.functions)
-
-  return indexes
-}
-
-export function getCategories(functions: VueUseFunction[]): string[] {
-  return uniq(
-    functions
-      .filter(i => !i.internal)
-      .map(i => i.category)
-      .filter(Boolean),
-  ).sort(
-    (a, b) => (a.startsWith('@') && !b.startsWith('@'))
-      ? 1
-      : (b.startsWith('@') && !a.startsWith('@'))
-        ? -1
-        : a.localeCompare(b),
-  )
 }
 
 export async function updateImport({ packages, functions }: PackageIndexes) {
